MetaphorMind-Cross-Metaphor-Detector-and-interpreter/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
transformers==4.36.2
torch==2.1.2
numpy==1.26.3
python-multipart==0.0.6
pydantic==2.5.3
googletrans==4.0.0-rc1
google-generativeai==0.8.5
google-genai==0.6.0
python-dotenv==1.0.0
langdetect==1.0.9
psutil==5.9.8
aiofiles==23.2.1
slowapi==0.1.9
redis==5.0.1
motor==3.3.2
pymongo==4.6.1
openai-whisper==20231117
requests==2.31.0
datasets==2.16.1
scikit-learn==1.3.2
evaluate==0.4.1
tenacity==8.2.3
bson==0.5.10
